+++ src/Modules.php
@@ -22,20 +22,28 @@ class Modules {
     $modules = (array) $modules;
     $enabled_modules = array();
     $t = get_t();
+    $enabled = $t('enabled');
+    $not_enabled = $t('not-enabled');
+    $report = array();
     // Check to see if each module is disabled.
     foreach ($modules as $module) {
       if (module_exists($module)) {
-        // This module is enabled, throw an exception.
-        $message = 'The module @module was supposed to be disabled by this update, but was not. Please investigate the problem and re-run this update.';
-        $variables = array('@module' => $module);
-        Message::make($message, $variables, WATCHDOG_ERROR);
-        throw new HudtException($message, $variables, WATCHDOG_ERROR, FALSE);
+        // This module is enabled.
+        $report[$module] = $enabled;
+      }
+      else {
+        $report[$module] = $not_enabled;
       }
     }
-    $module_list = implode(', ', $modules);
 
-    $message = "The modules @disabled are disabled.";
-    $variables = array('@disabled' => $module_list);
+    if (in_array($enabled, $report)) {
+      $message = 'The modules that were supposed to be disabled by this update, were not. Please investigate the problem and re-run this update. Report: !report';
+      $variables = array('!report' => $report);
+      throw new HudtException($message, $variables, WATCHDOG_ERROR, TRUE);
+    }
+
+    $message = "The requested modules were disabled. Report: !report";
+    $variables = array('!report' => $report);
     return Message::make($message, $variables, WATCHDOG_INFO);
   }
 
@@ -53,23 +61,93 @@ class Modules {
    */
   public static function checkEnabled($modules = array()) {
     $modules = (array) $modules;
-    $enabled_modules = array();
     $t = get_t();
+    $enabled = $t('enabled');
+    $not_enabled = $t('not-enabled');
+    $report = array();
     // Check to see if each module is enabled.
     foreach ($modules as $module) {
       if (!module_exists($module)) {
         // This module is not enabled, throw an exception.
-        $message = 'The module @module was supposed to be enabled by this update, but was not. Please investigate the problem and re-run this update.';
-        $variables = array('@module' => $module);
-        Message::make($message, $variables, WATCHDOG_ERROR);
-        throw new HudtException($message, $variables, WATCHDOG_ERROR, FALSE);
+        $report[$module] = $not_enabled;
       }
+      else {
+        $report[$module] = $enabled;
+      }
+    }
+
+    if (in_array($not_enabled, $report)) {
+      // Something was not enabled. Fail the update.
+      $message = 'Some of the modules that were supposed to be enabled, are not showing as enabled. Please investigate the problem and re-run this update. Report: !report';
+      $variables = array('!report' => $report);
+      throw new HudtException($message, $variables, WATCHDOG_ERROR, TRUE);
     }
     $module_list = implode(', ', $modules);
 
-    $message = "The modules @enabled were enabled successfully.";
-    $variables = array('@enabled' => $module_list);
+    $message = "The requested modules were enabled successfully. Report: !report";
+    $variables = array('!report' => $report);
     return Message::make($message, $variables, WATCHDOG_INFO);;
+  }
+
+
+  /**
+   * Checks for requested modules are present and optionally dependencies.
+   *
+   * @param array $modules
+   *   The module machine names to check for being present.
+   * @param bool $check_dependencies
+   *   TRUE [default] checks to see that module dependencies are present too.
+   *   FALSE does not check for dependencies
+   * @param array $module_data
+   *   Only used internally! The Drupal system data about existing modules.
+   * @param bool $failed_flag
+   *   Only used internally!  A persistant flag (by reference) to indicate a
+   *   required module is missing.
+   *
+   * @return array
+   *   A array of the report of any missing modules.
+   *
+   * @throws \HudtException
+   *   Calls the update a failure, preventing it from registering the update_N.
+   */
+  public static function checkPresent($modules = array(), $check_dependencies = TRUE, $module_data = NULL, &$failed_flag = FALSE) {
+    $modules = (array) $modules;
+    $report = array();
+    // If $module_data is NULL, then this is the first_run.
+    $first_run = (is_null($module_data)) ? TRUE : FALSE;
+    // Get all module data so we can see what we have, if we don't have it yet.
+    $module_data = (empty($module_data)) ? \system_rebuild_module_data() : $module_data;
+    // Create an associative array with weights as values.
+    foreach ($modules as $module) {
+      // Check for basic presence.
+      if (!isset($module_data[$module])) {
+        // This module is not found in the filesystem. Report it.
+        $report[$module] = t('missing');
+        $failed_flag = TRUE;
+      }
+      elseif ($check_dependencies === TRUE) {
+        // Check for dependencies, recursively.
+        if (!empty($module_data[$module]->requires)) {
+          $dependencies = array_keys($module_data[$module]->requires);
+          $missing_dependencies = self::checkPresent($dependencies, $check_dependencies, $module_data, $failed_flag);
+          // Only report it if there are missing dependencies.
+          if (!empty($missing_dependencies)) {
+            $report[$module]['dependency'] = $missing_dependencies;
+          }
+        }
+      }
+    }
+
+    if ($first_run && $failed_flag) {
+      $message = 'The check for presence of modules failed. Please investigate the problem and re-run this update.  Report: !report';
+      $variables = array(
+        '!report' => $report,
+      );
+      throw new HudtException($message, $variables, WATCHDOG_ERROR, TRUE);
+    }
+    else {
+      return $report;
+    }
   }
 
 
@@ -88,20 +166,42 @@ class Modules {
    *   Calls the update a failure, preventing it from registering the update_N.
    */
   public static function enable($modules = array(), $enable_dependencies = TRUE) {
-    $modules = (array) $modules;
-    $enable_good = module_enable($modules, $enable_dependencies);
-    if (!$enable_good) {
-      // Enable command failed.
-      $module_list = implode(', ', $modules);
+    try {
+      $modules = (array) $modules;
       $t = get_t();
-      $message = 'The requested modules @modules to be enabled by this update, were not, because one of them does not exist in the codebase. Please investigate the problem and re-run this update.';
-      $variables = array('@modules' => $module_list);
-      Message::make($message, $variables, WATCHDOG_ERROR);
-      throw new HudtException($message, $variables, WATCHDOG_ERROR, FALSE);
-    }
-    $success = self::checkEnabled($modules);
+      $enabled = $t('enabled');
+      $failed = $t('failed');
+      self::checkPresent($modules, $enable_dependencies);
+      $report = array();
+      foreach ($modules as $module) {
+        $enable_good = module_enable($module, $enable_dependencies);
+        if ($enable_good) {
+          $report[$module] = $enabled;
+        }
+        else {
+          $report[$module] = $failed;
+        }
+      }
+      $variables = array('!report' => $report);
 
-    return $success;
+      if (!in_array($failed, $report)) {
+        // Enable command failed.
+        $message = 'The modules to be enabled by this update, were not. Please investigate the problem and re-run this update. Report: !report';
+        throw new HudtException($message, $variables, WATCHDOG_ERROR, TRUE);
+      }
+      $success = self::checkEnabled($modules);
+      return $success;
+    }
+    catch (\Exception $e) {
+      $vars['!error'] = (method_exists($e, 'logMessage')) ? $e->logMessage() : $e->getMessage();
+
+      if (!method_exists($e, 'logMessage')) {
+        // Not logged yet, so log it.
+        $message = 'Modules::enable failed because: !error';
+        Message::make($message, $vars, WATCHDOG_ERROR);
+      }
+      throw new HudtException('Update aborted!  !error', $vars, WATCHDOG_ERROR, FALSE);
+    }
   }
 
 
@@ -117,14 +217,27 @@ class Modules {
    *   Messsage indicating the modules are disabled.
    */
   public static function disable($modules = array(), $disable_dependents = TRUE) {
-    $modules = (array) $modules;
-    module_disable($modules, $disable_dependents);
-    // Verify that the modules were disabled.
-    $success = self::checkDisabled($modules);
-    drupal_flush_all_caches();
-    $success .= Message::make("Caches cleared and Registry Rebuilt.", array(), WATCHDOG_INFO);
+    try {
+      $modules = (array) $modules;
+      $disable_dependents = !empty($disable_dependents) ? TRUE : FALSE;
+      module_disable($modules, $disable_dependents);
+      // Verify that the modules were disabled.
+      $success = self::checkDisabled($modules);
+      drupal_flush_all_caches();
+      $success .= Message::make("Caches cleared and Registry Rebuilt.", array(), WATCHDOG_INFO);
 
-    return $success;
+      return $success;
+    }
+    catch (\Exception $e) {
+      $vars['!error'] = (method_exists($e, 'logMessage')) ? $e->logMessage() : $e->getMessage();
+
+      if (!method_exists($e, 'logMessage')) {
+        // Not logged yet, so log it.
+        $message = 'Modules::disable failed because: !error';
+        Message::make($message, $vars, WATCHDOG_ERROR);
+      }
+      throw new HudtException('Caught Exception: Update aborted!  !error', $vars, WATCHDOG_ERROR, FALSE);
+    }
   }
 
 
@@ -143,35 +256,51 @@ class Modules {
    *   Calls the update a failure, preventing it from registering the update_N.
    */
   public static function uninstall($modules = array(), $uninstall_dependents = TRUE) {
-    $modules = (array) $modules;
-    $t = get_t();
-    module_disable($modules);
-    foreach ($modules as $module) {
-      if (module_exists($module)) {
-        // The module is not disabled, so it can not be uninstalled.
-        $variables = array('@module' => $module);
-        Message::make($message, $variables, WATCHDOG_ERROR);
+    try {
+      $modules = (array) $modules;
+      $uninstall_dependents = !empty($uninstall_dependents) ? TRUE : FALSE;
+      $t = get_t();
+      $uninstalled = $t('uninstalled');
+      $not_uninstalled = $t('not-uninstalled');
+      $enabled = $t('enabled');
+      $report = array();
+      foreach ($modules as $module) {
+        if (module_exists($module)) {
+          // The module is not disabled, so it can not be uninstalled.
+          $report[$module] = $enabled;
+        }
+        else {
+          // Made it this far. Safe to uninstall this module.
+          $success = drupal_uninstall_modules(array($module), $uninstall_dependents);
+          if ($success) {
+            $report[$module] = $uninstalled;
+          }
+          else {
+            $report[$module] = $not_uninstalled;
+          }
+        }
+      }
+
+      $variables = array('!report' => $report);
+      if (in_array($enabled, $report) || in_array($not_uninstalled, $report)) {
+        // Uninstalling the modules failed, can not be more specifc about why.
+        $message = "The modules requested to uninstall were NOT uninstalled successfully.  Report: !report";
         throw new HudtException($message, $variables, WATCHDOG_ERROR, FALSE);
       }
+      else {
+        $message = "The requested modules were uninstalled successfully. Report: !report";
+        return Message::make($message, $variables, WATCHDOG_INFO);
+      }
     }
-    // Made it this far.Safe to uninstall.
-    // Uninstall only the specified modules.  Do not uninstall dependents
-    // unless the are specified as well.
-    $uninstall_dependents = FALSE;
-    $success = drupal_uninstall_modules($modules, $uninstall_dependents);
+    catch (\Exception $e) {
+      $vars['!error'] = (method_exists($e, 'logMessage')) ? $e->logMessage() : $e->getMessage();
 
-    $module_list = implode(', ', $modules);
-    if ($success) {
-      $message = "The modules @uninstalled were uninstalled successfully.";
-      $variables = array('@uninstalled' => $module_list);
-      return Message::make($message, $variables, WATCHDOG_INFO);;
-    }
-    else {
-      // Uninstalling the modules failed, can not be more specifc about why.
-      $message = "The modules @uninstalled were NOT uninstalled successfully.  Check to see that the modules you are attempting to uninstall include any dependents in the correct order.";
-      $variables = array('@uninstalled' => $module_list);
-      Message::make($message, $variables, WATCHDOG_ERROR);
-      throw new HudtException($message, $variables, WATCHDOG_ERROR, FALSE);
+      if (!method_exists($e, 'logMessage')) {
+        // Not logged yet, so log it.
+        $message = 'Modules::uninstall failed because: !error';
+        Message::make($message, $vars, WATCHDOG_ERROR);
+      }
+      throw new HudtException('Caught Exception: Update aborted!  !error', $vars, WATCHDOG_ERROR, FALSE);
     }
   }